+++ shared/chat/conversation/messages/react-button/index.js
@@ -1,6 +1,5 @@
 // @flow
 import * as React from 'react'
-import * as ReactDOM from 'react-dom'
 import * as Types from '../../../../constants/types/chat2'
 import {
   Box2,
@@ -12,57 +11,9 @@ import {
   EmojiIfExists,
 } from '../../../../common-adapters'
 import * as Styles from '../../../../styles'
-import {Picker} from 'emoji-mart'
+import {Picker} from './picker'
 import {backgroundImageFn} from '../../../../common-adapters/emoji'
 import DelayInterval from './delay-interval'
-
-type WrapProps = {|
-  backgroundImageFn: (set: string, sheetSize: string) => string,
-  onClick: ({colons: string}, evt: Event) => void,
-|}
-
-class PickerWrapper extends React.Component<WrapProps> {
-  _picker: Picker
-
-  // Setting autoFocus={true} on Picker doesn't work, so focus it
-  // ourselves on mount/update.
-
-  componentDidMount = () => {
-    this._focus()
-  }
-
-  componentDidUpdate = () => {
-    this._focus()
-  }
-
-  _focus = () => {
-    if (!this._picker) {
-      return
-    }
-    const node = ReactDOM.findDOMNode(this._picker)
-    if (!node || !(node instanceof Element)) {
-      return
-    }
-    const input = node.querySelector('input')
-    if (!input) {
-      return
-    }
-    input.focus()
-  }
-
-  render = () => {
-    return (
-      <Picker
-        autoFocus={true}
-        emoji="star-struck"
-        ref={picker => (this._picker = picker)}
-        title="reacjibase"
-        onClick={this.props.onClick}
-        backgroundImageFn={this.props.backgroundImageFn}
-      />
-    )
-  }
-}
 
 export type Props = {|
   active: boolean,
@@ -265,7 +216,7 @@ export class NewReactionButton extends React.Component<NewReactionButtonProps, N
             position="top right"
             onHidden={() => this._setShowingPicker(false)}
           >
-            <PickerWrapper onClick={this._onAddReaction} backgroundImageFn={backgroundImageFn} />
+            <Picker onClick={this._onAddReaction} backgroundImageFn={backgroundImageFn} />
           </FloatingBox>
         )}
       </ButtonBox>
