+++ shared/chat/conversation/messages/react-button/picker.mobile.js
@@ -1,0 +1,12 @@
+// @flow
+import * as React from 'react'
+
+import {type Props} from './picker'
+
+class Picker extends React.Component<Props> {
+  render() {
+    return null
+  }
+}
+
+export {Picker}
